Master.videochat webrtc updated bug fixing roman

+++ sample-videochat-webrtc/src/main/java/com/quickblox/sample/groupchatwebrtc/services/CallService.java
@@ -190,9 +190,13 @@ public class CallService extends Service{
     }
 
     private void destroyRtcClientAndChat() {
-        rtcClient.destroy();
+        if(rtcClient != null) {
+            rtcClient.destroy();
+        }
         ChatPingAlarmManager.onDestroy();
-        chatService.destroy();
+        if(chatService != null) {
+            chatService.destroy();
+        }
         stopSelf();
     }
